Refactor binaries
