+++ README.md
@@ -47,8 +47,8 @@ Initialise the Elasticsearch index and start the server in read-write mode:
 
 ```bash
 export ES_HOST="..."
-./target/release/bookdb create-index
-./target/release/bookdb serve --allow-writes --upload-dir="<...>" <config file>
+./target/release/bookdb_ctl create-index
+./target/release/bookdb --allow-writes --upload-dir="<...>" <config file>
 ```
 
 Omit the `--alow-writes` to launch in read-only mode.
@@ -56,13 +56,13 @@ Omit the `--alow-writes` to launch in read-only mode.
 Dump the Elasticsearch index as json to stdout with:
 
 ```bash
-./target/release/bookdb export-index > bookdb.json
+./target/release/bookdb_ctl export-index > bookdb.json
 ```
 
 Restore it, overwriting the existing index:
 
 ```bash
-./targets/release/bookdb import-index --drop-existing < bookdb.json
+./targets/release/bookdb_ctl import-index --drop-existing < bookdb.json
 ```
 
 See the `--help` text for more.
